CRAB-17 repo structure changes
