+++ docker/Dockerfile
@@ -1,0 +1,60 @@
+ARG  BUILDER_IMAGE=golang:alpine
+############################
+# STEP 1 build executable binary
+############################
+FROM ${BUILDER_IMAGE} as builder
+
+# Install git + SSL ca certificates.
+# Git is required for fetching the dependencies.
+# Ca-certificates is required to call HTTPS endpoints.
+RUN apk update && apk add --no-cache git ca-certificates tzdata && update-ca-certificates
+
+# Create appuser
+ENV USER=msuser
+ENV UID=2115
+
+ARG MSNAME
+
+# See https://stackoverflow.com/a/55757473/12429735
+RUN adduser \
+  --disabled-password \
+  --gecos "" \
+  --home "/nonexistent" \
+  --shell "/sbin/nologin" \
+  --no-create-home \
+  --uid "${UID}" \
+  "${USER}"
+
+WORKDIR $GOPATH/src/
+
+COPY ./$MSNAME/ .
+COPY ./go.mod .
+COPY ./go.sum .
+
+
+RUN go mod tidy && go get
+
+# Build the binary
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=arm64 go build \
+  -ldflags='-w -s -extldflags "-static"' -a \
+  -o $GOPATH/src/bin/ $GOPATH/src/server
+
+############################
+# STEP 2 build a small image
+############################
+FROM scratch
+
+# Import from builder.
+COPY --from=builder /usr/share/zoneinfo /usr/share/zoneinfo
+COPY --from=builder /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/
+COPY --from=builder /etc/passwd /etc/passwd
+COPY --from=builder /etc/group /etc/group
+
+# Copy our static executable
+COPY --from=builder $GOPATH/src/bin/ /go/bin/exec
+
+# Use an unprivileged user.
+USER msuser:msuser
+
+# Run the exec binary.
+ENTRYPOINT ["/go/bin/exec"]
