+++ services/api/go.mod
@@ -1,0 +1,3 @@
+module github.com/CrabStash/crab-stash/api
+
+go 1.18
